requirements.txt
blinker==1.6.2
click==8.1.3
colorama==0.4.6
distlib==0.3.6
EasyProcess==1.1
entrypoint2==1.1
filelock==3.12.2
Flask==2.3.2
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
MouseInfo==0.1.3
mss==9.0.1
Pillow==10.0.0
platformdirs==3.8.1
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
pyscreenshot==3.1
PyScreeze==0.1.29
pytweening==1.0.7
subprocess.run==0.0.8
virtualenv==20.23.1
Werkzeug==2.3.6
